Updated all the typescripts to 4.8.3 Moved const dictionary declaration from types package to frontend Disabled some unused code that now doesn't compile with new typescript
