+++ application/common/elsa-constants/package.json
@@ -8,6 +8,6 @@
   "license": "MIT",
   "dependencies": {},
   "devDependencies": {
-    "typescript": "4.8.2"
+    "typescript": "4.8.3"
   }
 }
