+++ application/common/elsa-types/package.json
@@ -7,9 +7,9 @@
   "author": "Andrew Patterson",
   "license": "MIT",
   "dependencies": {
-    "@sinclair/typebox": "~0.24.20"
+    "@sinclair/typebox": "0.24.41"
   },
   "devDependencies": {
-    "typescript": "~4.7.4"
+    "typescript": "4.8.3"
   }
 }
