+++ application/common/elsa-types/schemas-duo.ts
@@ -40,47 +40,13 @@ const KnownLimitationCodes = [
   "DUO:0000004",
 ] as const;
 
-const KnownDuoCodes = [
-  ...KnownModifierCodes,
-  ...KnownLimitationCodes,
-] as const;
+const KnownDuoCodes = [...KnownModifierCodes, ...KnownLimitationCodes] as const;
 
 export type KnownModifierCode = typeof KnownModifierCodes[number];
 
 export type KnownLimitationCode = typeof KnownLimitationCodes[number];
 
 export type KnownDuoCode = typeof KnownDuoCodes[number];
-
-export function isKnownDuoCode(x: any): x is KnownDuoCode {
-  return x in KnownDuoCodes;
-}
-
-export type CodeToDescriptionType = {
-  [key in KnownDuoCode]: string;
-};
-
-export const duoCodeToDescription: CodeToDescriptionType = {
-  "DUO:0000004": "No restriction",
-  "DUO:0000006": "Health or medical or biomedical research",
-  "DUO:0000007": "Disease specific research",
-  "DUO:0000011": "Population origins or ancestry research only",
-  "DUO:0000015": "No general methods research",
-  "DUO:0000016": "Genetic studies only",
-  "DUO:0000018": "Not for profit",
-  "DUO:0000019": "Publication required",
-  "DUO:0000020": "Collaboration required",
-  "DUO:0000022": "Geographical restriction",
-  "DUO:0000024": "Publication moratorium",
-  "DUO:0000025": "Time limit on use",
-  "DUO:0000026": "User specific restriction",
-  "DUO:0000027": "Project specific restriction",
-  "DUO:0000028": "Institution specific restriction",
-  "DUO:0000029": "Return to database or resource",
-  "DUO:0000042": "General research use",
-  "DUO:0000043": "Clinical care use",
-  "DUO:0000045": "Not for profit organisation use only",
-  "DUO:0000046": "Non-commercial use only",
-};
 
 export const DuoPublicationRequiredSchema = Type.Object({
   code: Type.Literal("DUO:0000019"), // PUB
