+++ application/frontend/package.json
@@ -64,11 +64,11 @@
     "@types/react-table": "^7.7.12",
     "autoprefixer": "^10.4.7",
     "babel-plugin-macros": "^3.1.0",
-    "postcss": "8.4.14",
-    "prettier": "2.6.2",
+    "postcss": "8.4.16",
+    "prettier": "2.7.1",
     "react-scripts": "5.0.1",
-    "tailwindcss": "3.0.24",
-    "typescript": "4.6.4"
+    "tailwindcss": "3.1.8",
+    "typescript": "4.8.3"
   },
   "browserslist": {
     "production": [
