+++ application/frontend/src/components/rh/rh-form-with-save-button.tsx
@@ -1,21 +1,17 @@
 import React, { PropsWithChildren } from "react";
-import {
-  DeepPartial,
-  SubmitHandler,
-  UnpackNestedValue,
-  useForm,
-  UseFormReturn,
-} from "react-hook-form";
+import { SubmitHandler, UseFormReturn } from "react-hook-form";
 
+/*
 type Props<TFormValues> = {
   onSubmit: SubmitHandler<TFormValues>;
   methods: UseFormReturn<TFormValues>;
   children: (methods: UseFormReturn<TFormValues>) => React.ReactNode;
 };
 
-/**
- * A react-hook-form that contains other form inputs and a save button
- */
+// disabled due to compile error with newer Typescript
+// but we aren't currently doing Form input this way so deferring until we maybe
+// have complex forms
+
 export const RhFormWithSaveButton = <TFormValues extends Record<string, any>>({
   onSubmit,
   methods,
@@ -39,3 +35,4 @@ export const RhFormWithSaveButton = <TFormValues extends Record<string, any>>({
     </form>
   );
 };
+*/
